refactor: create hook to get active payouts
