+++ src/components/PayoutList/index.tsx
@@ -1,4 +1,4 @@
-import React, { useMemo, useState, useEffect } from "react";
+import React, { useMemo } from "react";
 
 import { TableBody } from "@material-ui/core";
 import { ColonyRole } from "@colony/colony-js";
@@ -8,20 +8,12 @@ import NewPayoutRow from "./NewPayoutRow";
 import Table from "../common/StyledTable";
 
 import { useSafeInfo } from "../../contexts/SafeContext";
-import { useColonyClient, useHasDomainPermission } from "../../contexts/ColonyContext";
-
-import getActivePayouts from "../../utils/colony/getColonyPayouts";
-import { PayoutInfo } from "../../typings";
+import { useActivePayouts, useHasDomainPermission } from "../../contexts/ColonyContext";
 
 const PayoutList = () => {
   const safeInfo = useSafeInfo();
-  const colonyClient = useColonyClient();
+  const activePayouts = useActivePayouts();
   const hasRootPermission = useHasDomainPermission(safeInfo?.safeAddress, 1, ColonyRole.Root);
-  const [activePayouts, setActivePayouts] = useState<PayoutInfo[]>([]);
-
-  useEffect(() => {
-    if (colonyClient) getActivePayouts(colonyClient).then(setActivePayouts);
-  }, [colonyClient]);
 
   const payoutList = useMemo(
     () => activePayouts.map(payout => <PayoutRow key={payout.blockTimestamp.toString()} payout={payout} />),
