+++ src/contexts/ColonyContext/hooks.ts
@@ -5,11 +5,12 @@ import { TokenInfo } from "@colony/colony-js/lib/clients/TokenClient";
 import { BigNumber, BigNumberish } from "ethers/utils";
 import { getPermissionProofs } from "@colony/colony-js/lib/clients/Colony/extensions/commonExtensions";
 import { Zero } from "ethers/constants";
-import { Domain, PermissionProof, MoveFundsBetweenPotsProof, Token } from "../../typings";
+import { Domain, PermissionProof, MoveFundsBetweenPotsProof, Token, PayoutInfo } from "../../typings";
 import userHasDomainRole from "../../utils/colony/userHasDomainRole";
 import { MAX_U256 } from "../../constants";
 import { useColonyContext } from "./ColonyContext";
 import getDomainTokenBalance from "../../utils/colony/getDomainTokenBalance";
+import getActivePayouts from "../../utils/colony/getColonyPayouts";
 
 export const useColonyClient = (): ColonyClient | undefined => {
   const { colonyClient } = useColonyContext();
@@ -204,4 +205,15 @@ export const useDomainTokenBalance = (domainId: BigNumberish, token: string): Bi
   }, [colonyClient, colonyDomains, domainId, token]);
 
   return balance;
+};
+
+export const useActivePayouts = (): PayoutInfo[] => {
+  const colonyClient = useColonyClient();
+  const [activePayouts, setActivePayouts] = useState<PayoutInfo[]>([]);
+
+  useEffect(() => {
+    if (colonyClient) getActivePayouts(colonyClient).then(setActivePayouts);
+  }, [colonyClient]);
+
+  return activePayouts;
 };
